move key material to storage base
