+++ packages/fireproof/test/valet.test.js
@@ -26,7 +26,7 @@ describe('new Valet', () => {
         assert('car parked')
         return val.getValetBlock('bafyreieth2ckopwivda5mf6vu76xwqvox3q5wsaxgbmxy2dgrd4hfuzmma').then((block) => {
           assert.equal(block.length, 300)
-          return val.getCarCIDForCID('bafyreieth2ckopwivda5mf6vu76xwqvox3q5wsaxgbmxy2dgrd4hfuzmma').then(({ result }) => {
+          return val.loader.getCarCIDForCID('bafyreieth2ckopwivda5mf6vu76xwqvox3q5wsaxgbmxy2dgrd4hfuzmma').then(({ result }) => {
             assert.equal(result, 'carCid')
           })
         })
